Simplify away per piece king ring attack count (#361)
Passed nonregr STC
```
Test  | simplring
Elo   | 2.15 +- 2.63 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [-3.00, 0.00]
Games | N: 22980 W: 5882 L: 5740 D: 11358
Penta | [271, 2744, 5367, 2788, 320]
```
https://ob.cwchess.org/test/1164/

Passed nonregr LTC
```
Test  | simplring
Elo   | 2.16 +- 2.49 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=64MB
LLR   | 2.94 (-2.94, 2.94) [-3.00, 0.00]
Games | N: 21416 W: 5195 L: 5062 D: 11159
Penta | [120, 2522, 5287, 2663, 116]
```
https://ob.cwchess.org/test/1165/

Bench: 15772962

+++ src/eval_types.hpp
@@ -179,11 +179,11 @@ public:
 #else
 class TunableSigmoid {
 private:
-    static constexpr i32 TABLE_SIZE = 256;
-    static constexpr i32 FP_SHIFT   = 16;
-    static constexpr i32 FP_ONE     = 1 << FP_SHIFT;
-    static constexpr f64 B          = static_cast<f64>(B_SCALE);
-    static constexpr f64 LN_39      = 3.6635616461296463;
+    static constexpr usize TABLE_SIZE = 256;
+    static constexpr i32   FP_SHIFT   = 16;
+    static constexpr i32   FP_ONE     = 1 << FP_SHIFT;
+    static constexpr f64   B          = static_cast<f64>(B_SCALE);
+    static constexpr f64   LN_39      = 3.6635616461296463;
 
     struct Table {
         i32                         range_min;
@@ -216,7 +216,7 @@ private:
         tbl.scale_fp =
           static_cast<i32>((static_cast<i64>(TABLE_SIZE - 1) << FP_SHIFT) / tbl.range_span);
 
-        for (i32 i = 0; i < TABLE_SIZE; ++i) {
+        for (usize i = 0; i < TABLE_SIZE; ++i) {
             const f64 t   = static_cast<f64>(i) / (TABLE_SIZE - 1);
             const f64 x   = tbl.range_min + t * tbl.range_span;
             const f64 z   = (x + c) / B;
@@ -230,8 +230,8 @@ private:
         const i32 x      = std::clamp(static_cast<i32>(x_val), tbl.range_min, tbl.range_max);
         const i64 idx_fp = static_cast<i64>(x - tbl.range_min) * tbl.scale_fp;
 
-        const i32 idx  = static_cast<i32>(idx_fp >> FP_SHIFT);
-        const i32 frac = static_cast<i32>(idx_fp & (FP_ONE - 1));
+        const usize idx  = static_cast<usize>(idx_fp >> FP_SHIFT);
+        const i32   frac = static_cast<i32>(idx_fp & (FP_ONE - 1));
 
         const i32 v0 = tbl.values[idx];
         const i32 v1 = tbl.values[std::min(idx + 1, TABLE_SIZE - 1)];
